# openapi.yaml
openapi: 3.0.0
info:
  title: SpaceShare API — Full Live Coverage
  description: >-
    Every endpoint actually mounted in app.js as of the final patch, verified by
    booting the real server against a live database. Endpoints that exist in
    code but aren't reachable (e.g. the old unmounted payment module) are
    excluded.
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Root Server
  - url: http://localhost:4000/api
    description: API Base Server
paths:
...
  /workspaces:
    post:
      summary: Create Workspace (Host)
      tags:
        - Workspaces
      parameters: []
      responses:
        '200':
          description: Successful response
      requestBody:
        content:
          application/json:
            example:
              title: Quiet Desk Lagos
              workspace_type: desk
              capacity: 2
              city: Lagos
              address: 12 Adeola Odeku St
              latitude: 6.4281
              longitude: 3.4219
      security:
        - bearerAuth: []
    get:
      summary: List Workspaces (filters)
      tags:
        - Workspaces
      parameters:
        - name: city
          in: query
          required: false
          schema:
            type: string
        - name: minCapacity
          in: query
          required: false
          schema:
            type: string
        - name: minPrice
          in: query
          required: false
          schema:
            type: string
        - name: maxPrice
          in: query
          required: false
          schema:
            type: string
        - name: amenities
          in: query
          required: false
          schema:
            type: string
        - name: minReliabilityScore
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
  /workspaces/search:
    get:
      summary: Search Workspaces
      tags:
        - Workspaces
      parameters:
        - name: latitude
          in: query
          required: false
          schema:
            type: string
        - name: longitude
          in: query
          required: false
          schema:
            type: string
        - name: radius
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
...
  /favorites:
    get:
      summary: List Favorites
      tags:
        - Favorites
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
        - bearerAuth: []
  /bookings:
    post:
      summary: Create Booking (Seeker)
      tags:
        - Bookings
      parameters: []
      responses:
        '200':
          description: Successful response
      requestBody:
        content:
          application/json:
            example:
              workspace_id: '{{workspaceId}}'
              start_time: '2026-08-20T09:00:00Z'
              end_time: '2026-08-20T12:00:00Z'
      security:
        - bearerAuth: []
    get:
      summary: List My Bookings
      tags:
        - Bookings
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
        - bearerAuth: []
...
  /admin/bookings:
    get:
      summary: Get All Bookings
      tags:
        - Admin
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT